Move args to its own file
