+++ src/engine/profits.ts
@@ -13,7 +13,7 @@ import {
   toInt,
 } from "kolmafia";
 import { $item, $items, get, getSaleValue, Session, set, sumNumbers } from "libram";
-import { args } from "../main";
+import { args } from "../args";
 
 function currency(...items: Item[]): () => number {
   const unitCost: [Item, number][] = items.map((i) => {
