+++ src/tasks/utils.ts
@@ -33,7 +33,7 @@ import {
   Snapper,
 } from "libram";
 import { garboAverageValue, garboValue } from "../engine/profits";
-import { args } from "../main";
+import { args } from "../args";
 
 export function haveAll(its: Item[]): boolean {
   return its.reduce((a, it) => a && have(it), true);
